feat: 항목 삭제 시 컨펌창

+++ src/components/TodoList/Item.tsx
@@ -20,7 +20,9 @@ const Item = ({item, updateTodo, deleteTodo, modifyModeId, setModifyModeId}: Ite
     };
 
     const handleDeleteClick = () => {
-        deleteTodo(id);
+        if (window.confirm(`'${todo}' 항목을 삭제하시겠습니까?`)) {
+            deleteTodo(id);
+        }
     };
 
     const handleModifyClick = () => {
